formation-fhir-python/requirements.txt
iris-pex-embedded-python
google-auth-oauthlib==1.0.0
fhir.resources
intersystems_iris @ https://github.com/intersystems-community/intersystems-irispython/releases/download/3.7.6/intersystems_iris-3.7.6-py3-none-any.whl
PyJWT
fhir.resources
pytest
deepdiff
